Revert "Added more unit tests"
