+++ autoprof/image/target_image.py
@@ -168,7 +168,7 @@ class Target_Image(BaseImage):
             pixelscale = self.pixelscale * scale,
             zeropoint = self.zeropoint,
             variance = self.variance[:MS*scale, :NS*scale].reshape(MS, scale, NS, scale).sum(axis=(1, 3)) if self.has_variance else None,
-            mask = self.mask[:MS*scale, :NS*scale].reshape(MS, scale, NS, scale).amax(axis=(1, 3)) if self.has_mask else None,
+            mask = self.mask[:MS*scale, :NS*scale].reshape(MS, scale, NS, scale).max(axis=(1, 3)) if self.has_mask else None,
             psf = self.psf[:PMS*scale, :PNS*scale].reshape(PMS, scale, PNS, scale).sum(axis=(1, 3)) if self.has_psf else None,
             psf_upscale = self.psf_upscale,
             note = self.note,
